Make floating IPs optional
At the moment, CAPO expects the control plane address to be a floating
IP. We're going to make it configurable through the "useControlPlaneFIP"
option.

In order to remain backwards compatible, it will default to "true".
If set to "false", we'll just use the control plane IP address without
attempting to create/attach a floating IP.